web-client/task-creation: description text
