+++ web-client/src/components/simple/DISCOllaborative.vue
@@ -1,0 +1,5 @@
+<template>
+  <span>
+    <span class="font-disco text-disco-cyan">DIS</span>
+    <span class="font-disco text-disco-blue">CO</span>llaborative</span>
+</template>
